Added Gamestatesystem.
As an example, a pausestate has been implemented.
By pressing 'p' it is now possible to enter and
leave the pausescreen (which at the moment is just
a black screen). All gamestates must inherit from
Gamestate and implement its methods. The transitions
can be implemented by setState, a method from Game.